requirements.txt
asgiref
bleach
Django
django-jazzmin
django-summernote
django-taggit
sqlparse
tzdata
webencodings
